+++ grafting/src/raw_gl/windows.rs
@@ -227,9 +227,8 @@ pub fn import_gl_framebuffer_vulkan_win32(
         gl_with_extensions.DeleteMemoryObjectsEXT(1, &memory_object);
 
         // Wrap the Vulkan image as a wgpu texture
-        let imported = host
-            .device
-            .create_texture_from_hal::<wgpu::wgc::api::Vulkan>(
+        let imported = crate::wgpu_compat::create_texture_from_hal::<wgpu_hal::api::Vulkan>(
+                &host.device,
                 hal_device.texture_from_raw(
                     vulkan_image,
                     &wgpu_hal::TextureDescriptor {