+++ grafting/src/surfman_gl/windows_dx12_shared.rs
@@ -411,7 +411,8 @@ fn init_size_dependent_state(
             1,
         );
 
-        let wgpu_texture = wgpu_device.create_texture_from_hal::<wgpu::wgc::api::Dx12>(
+        let wgpu_texture = crate::wgpu_compat::create_texture_from_hal::<wgpu_hal::api::Dx12>(
+                &wgpu_device,
             hal_texture,
             &wgpu::TextureDescriptor {
                 label: Some("angle-d3d11-shared-dx12-import"),